kirsch_portfolio_project/requirements.txt
Django>=3.2,<4.0
psycopg2-binary>=2.9,<3.0
Pillow>=8.0,<9.0
djangorestframework
django-cors-headers